server/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django
django-cors-headers==4.6.0
django-cron
djangorestframework==3.15.2
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
requests
bs4
django-redis
django-crontab
cryptography